nk3: Move provisioner into trussed module
